+++ src/crawl/revisit-policy.spec.ts
@@ -1,0 +1,11 @@
+import { mock } from 'jest-mock-extended';
+
+import { defaultRevisitPolicy } from './revisit-policy';
+
+describe('defaultRevisitPolicy', () => {
+  describe('canRevisit', () => {
+    it('should return false', () => {
+      expect(defaultRevisitPolicy.canRevisit(mock())).toBe(false);
+    });
+  });
+});