+++ src/index.ts
@@ -1,11 +1,12 @@
 import 'reflect-metadata';
 
+export * from './crawl';
 export * from './follow';
-export * from './href';
 export * from './models';
 export * from './parse';
 export * from './resolve';
 export * from './serialize';
 export * from './submit';
+export * from './utils';
 export * from './visitor';
 export * from './validate';